[php5.x] Improve behat tests with type hints. Fix bugs found. Introduce getter/setter method generation on BaseEntity classes.

+++ tests/Unit/Dms/MySQL/Generator/Dms/DmsColumnTest.php
@@ -160,12 +160,12 @@ class DmsColumnTest extends TestCase
             [
                 1,
                 'int',
-                'integer',
+                'int',
             ],
             [
                 1.1,
                 'float',
-                'double',
+                'float',
             ],
             [
                 '1',
@@ -291,13 +291,13 @@ class DmsColumnTest extends TestCase
                     'timestamp',
                     'tinyint',
                 ],
-                'integer'
+                'int'
             ],
             [
                 [
                     'float',
                 ],
-                'double',
+                'float',
             ],
             [
                 [
